[artf791857][artf765279][COVERITY]: Fix code defects generated by coverity tool.
1. CID:59988 Concurrent data access voilation.
2. CID:62279 Invalid assignment with Unsigned values.
3. CID:63166 Do not use int inside for loop
4. CID:75798 Constant expression result

+++ SN100x/halimpl/tml/phOsalNfc_Timer.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2010-2014 NXP Semiconductors
+ * Copyright 2010-2014, 2020 NXP
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -314,7 +314,7 @@ void phOsalNfc_Timer_Cleanup(void) {
 *******************************************************************************/
 static void phOsalNfc_DeferredCall(void* pParams) {
   /* Retrieve the timer id from the parameter */
-  uint32_t dwIndex;
+  unsigned long dwIndex;
   phOsalNfc_TimerHandle_t* pTimerHandle;
   if (NULL != pParams) {
     /* Retrieve the index at which the timer handle structure is stored */
